refactor: centralize config import
