+++ deploy-commands.js
@@ -1,5 +1,5 @@
 const { REST, Routes, SlashCommandBuilder } = require('discord.js');
-const { clientId, guildId, token } = require('./config.json');
+const { clientId, guildId, token } = require('./config');
 const fs = require('node:fs');
 const path = require('node:path');
 const dbm = require('./database-manager');
